requirements.txt
rumps
pillow
pyobjc-framework-Quartz
pynput